todo: finish refactor into editor_buffer.rs module
