+++ src/render.rs
@@ -126,76 +126,6 @@ impl WidgetBufferCoordTransformation {
     }
 }
 
-#[derive(QueryData)]
-#[query_data(mutable)]
-struct EditorBuffer {
-    editor: Option<&'static mut CosmicEditor>,
-    buffer: &'static mut CosmicEditBuffer,
-}
-
-impl EditorBufferItem<'_> {
-    fn with_buffer_mut<F: FnOnce(&mut Buffer) -> T, T>(&mut self, f: F) -> T {
-        match self.editor.as_mut() {
-            Some(editor) => editor.with_buffer_mut(f),
-            None => f(&mut self.buffer.0),
-        }
-    }
-
-    fn with_buffer<F: FnOnce(&Buffer) -> T, T>(&self, f: F) -> T {
-        match self.editor.as_ref() {
-            Some(editor) => editor.with_buffer(f),
-            None => f(&self.buffer.0),
-        }
-    }
-
-    fn borrow_with<'a>(
-        &'a mut self,
-        font_system: &'a mut cosmic_text::FontSystem,
-    ) -> ManuallyBorrowedWithFontSystem<'a, Self> {
-        ManuallyBorrowedWithFontSystem {
-            font_system,
-            inner: self,
-        }
-    }
-}
-
-pub struct ManuallyBorrowedWithFontSystem<'a, T> {
-    font_system: &'a mut cosmic_text::FontSystem,
-    inner: &'a mut T,
-}
-
-impl ManuallyBorrowedWithFontSystem<'_, EditorBufferItem<'_>> {
-    pub fn with_buffer_mut<F: FnOnce(&mut cosmic_text::BorrowedWithFontSystem<Buffer>) -> T, T>(
-        &mut self,
-        f: F,
-    ) -> T {
-        match self.inner.editor.as_mut() {
-            Some(editor) => editor.borrow_with(self.font_system).with_buffer_mut(f),
-            None => f(&mut self.inner.buffer.borrow_with(self.font_system)),
-        }
-    }
-}
-
-impl BufferMutExtras for ManuallyBorrowedWithFontSystem<'_, EditorBufferItem<'_>> {
-    fn width(&mut self) -> f32 {
-        self.with_buffer_mut(|b| b.width())
-    }
-
-    fn height(&mut self) -> f32 {
-        self.with_buffer_mut(|b| b.height())
-    }
-
-    fn compute(&mut self) {
-        self.with_buffer_mut(|b| b.compute())
-    }
-}
-
-// impl BufferMutExtras for BorrowedWithFontSystem<'_, EditorOrBufferItem<'_>> {
-//     fn width(&mut self) -> f32 {
-
-//     }
-// }
-
 /// Renders to the [CosmicRenderOutput]
 fn render_texture(
     mut query: Query<(
